fix: Ignore remote peer closed on write (ConnectionReset)
During writing to a connection stream a socket exception (SocketException) may be thrown.

The socket error SocketError.ConnectionReset (10054 WSAECONNRESET) indicates that the remote peer closed the connection.

This lead to unhandled exceptions being thrown and logged.

Given that this is a client initiated abort, there is no issue on the server side.
A debug log level log message is appropriate for that, instead of warning or throwing (unhandled) exceptions.

---

This is reproducible with mass-connects.
In my test scenario I create (sequentially) 20 times (concurrently) 400 TCP connect Tasks.

Co-authored-by: Jan Klass <jan.klass@ite-si.de>